pyproject.toml
[tool.poetry]
name = "ReferralSystem"
version = "0.1.0"
authors = ["You"]



[tool.poetry.dependencies]
python = "^3.12"
asyncpg = "^0.29.0"
fastapi = "^0.111.0"
uvicorn = { extras = ["standard"], version = "^0.30.0" }
sqlalchemy = "^2.0"
python-dotenv = "^1.0.1"
aiogram = "3.22.0"
jinja2 = "^3.1.3"

itsdangerous = "^2.2.0"
python-multipart = "^0.0.9"
psycopg2 = "^2.9.11"
pydantic-settings = "^2.12.0"
alembic = "^1.17.2"
aiosqlite = "^0.21.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-asyncio = "^0.23"
pytest-mock = "^3.12"
black = "^24.8"
flake8 = "^7.1"
mypy = "^1.10"
isort = "^5.13"
pre-commit = "^3.8"
sqlalchemy2-stubs = "0.0.2a38"
httpx = "^0.28.1"

[tool.black]
line-length = 79
target-version = ['py310']
skip-string-normalization = true

[tool.ruff]
select = ["ALL"]
ignore = [
    "E501",
]
line-length = 120
fix = true

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.isort]
profile = "black"
line_length = 79
multi_line_output = 3
include_trailing_comma = true
combine_as_imports = true
force_grid_wrap = 0

[tool.flake8]
max-line-length = 79
exclude = ["venv", ".venv", "build", "dist", "__pycache__"]
extend-ignore = ["E203", "W503"]

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
strict = false
warn_unused_ignores = true
warn_return_any = false
warn_unreachable = false
disable_error_code = ["attr-defined", "misc", "call-arg", "assignment"]

[[tool.mypy.overrides]]
module = "sqlalchemy.*"
ignore_missing_imports = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
